refactor(web): 清理死代码、拆分巨型组件、治理静默吞错
- 删除 @radix-ui/react-slot 及确证死代码
- chat-messages 659→297、interactive-renderers 650→141、core-renderers 563→347 等 5 组件拆分
- 27 处 .catch(()=>{}) → 统一 ignoreError；确认框 ref/state 合一
- sse-handler execution-step dispatch 去重
- web 267 测试全过
