+++ apps/web/src/lib/catalog/core-renderers.tsx
@@ -1,16 +1,36 @@
-import { Children, useState, type ReactNode } from "react";
-import { useTranslation } from "react-i18next";
+/**
+ * core-renderers — public re-export barrel.
+ *
+ * Layout + display renderers live here (Stack, Row, Grid, Separator,
+ * Text, Badge, Icon, TagList, StatBar, Progress, Accordion, Section,
+ * renderJsonValue, JsonView).
+ *
+ * Card-family renderers (Card, CardList, EntryCard) are in card-renderers.tsx.
+ * Shared color/icon lookup maps are in catalog-constants.ts.
+ * Collapse state logic is in use-collapsible.ts.
+ *
+ * All symbols remain importable from this path so existing consumers
+ * continue to work without change.
+ */
+
+import { useState, type ReactNode } from "react";
 import type { ComponentRenderer } from "@json-render/react";
 import { clsx } from "clsx";
 import * as Icons from "lucide-react";
 import { resolveIcon, useI18nResolver } from "./helpers.js";
+import {
+  gapClasses,
+  alignClasses,
+  justifyClasses,
+  iconSizeClasses,
+  badgeColorMap,
+} from "./catalog-constants.js";
 
 // ── Layout Components ────────────────────────────────────────────
 
 export const Stack: ComponentRenderer = ({ element, children }) => {
   const gap = (element.props?.gap as string) ?? "md";
-  const gapClass =
-    { xs: "gap-1", sm: "gap-2", md: "gap-3", lg: "gap-4" }[gap] ?? "gap-3";
+  const gapClass = gapClasses[gap] ?? "gap-3";
   return <div className={clsx("flex flex-col", gapClass)}>{children}</div>;
 };
 
@@ -18,20 +38,9 @@ export const Row: ComponentRenderer = ({ element, children }) => {
   const gap = (element.props?.gap as string) ?? "sm";
   const align = (element.props?.align as string) ?? "center";
   const justify = element.props?.justify as string | undefined;
-  const gapClass =
-    { xs: "gap-1", sm: "gap-2", md: "gap-3", lg: "gap-4" }[gap] ?? "gap-2";
-  const alignClass =
-    { start: "items-start", center: "items-center", end: "items-end" }[align] ??
-    "items-center";
-  const justifyClass = justify
-    ? ({
-        start: "justify-start",
-        center: "justify-center",
-        end: "justify-end",
-        between: "justify-between",
-        around: "justify-around",
-      }[justify] ?? "")
-    : "";
+  const gapClass = gapClasses[gap] ?? "gap-2";
+  const alignClass = alignClasses[align] ?? "items-center";
+  const justifyClass = justify ? (justifyClasses[justify] ?? "") : "";
   return (
     <div className={clsx("flex flex-row", gapClass, alignClass, justifyClass)}>
       {children}
@@ -92,19 +101,12 @@ export const Badge: ComponentRenderer = ({ element }) => {
   const resolve = useI18nResolver();
   const label = resolve(element.props?.label);
   const color = element.props?.color as string;
-  const colorMap: Record<string, string> = {
-    red: "bg-red-500/10 text-red-600 border-red-500/30",
-    amber: "bg-amber-500/10 text-amber-600 border-amber-500/30",
-    blue: "bg-blue-500/10 text-blue-600 border-blue-500/30",
-    green: "bg-green-500/10 text-green-600 border-green-500/30",
-    purple: "bg-purple-500/10 text-purple-600 border-purple-500/30",
-    cyan: "bg-cyan-500/10 text-cyan-600 border-cyan-500/30",
-  };
   return (
     <span
       className={clsx(
         "ui-chip inline-flex items-center px-1.5 py-0.5 text-[10px] border",
-        colorMap[color ?? ""] ?? "bg-muted text-muted-foreground border-border",
+        badgeColorMap[color ?? ""] ??
+          "bg-muted text-muted-foreground border-border",
       )}
     >
       {label}
@@ -117,9 +119,7 @@ export const Icon: ComponentRenderer = ({ element }) => {
   const size = (element.props?.size as string) ?? "sm";
   const LucideIcon = resolveIcon(name);
   if (!LucideIcon) return null;
-  const sizeClass =
-    { xs: "w-3 h-3", sm: "w-4 h-4", md: "w-5 h-5", lg: "w-6 h-6" }[size] ??
-    "w-4 h-4";
+  const sizeClass = iconSizeClasses[size] ?? "w-4 h-4";
   return <LucideIcon className={sizeClass} />;
 };
 
@@ -141,224 +141,6 @@ export const TagList: ComponentRenderer = ({ element }) => {
 };
 
 // ── Data Components ──────────────────────────────────────────────
-
-export const Card: ComponentRenderer = ({ element, children }) => {
-  const variant = element.props?.variant as string;
-  const collapsible = (element.props?.collapsible as boolean) ?? false;
-  const defaultExpanded = (element.props?.defaultExpanded as boolean) ?? false;
-  const [expanded, setExpanded] = useState(defaultExpanded);
-
-  let body: ReactNode = children;
-  if (collapsible) {
-    const items = Children.toArray(children);
-    const head = items[0];
-    const rest = items.slice(1);
-    const Chevron = expanded ? Icons.ChevronDown : Icons.ChevronRight;
-    body = (
-      <>
-        <div
-          className="flex items-center gap-2 cursor-pointer select-none"
-          onClick={() => setExpanded((v) => !v)}
-          role="button"
-          tabIndex={0}
-          aria-expanded={expanded}
-          onKeyDown={(e) => {
-            if (e.key === "Enter" || e.key === " ") {
-              e.preventDefault();
-              setExpanded((v) => !v);
-            }
-          }}
-        >
-          <Chevron
-            className="w-3 h-3 shrink-0 text-muted-foreground"
-            aria-hidden="true"
-          />
-          <div className="flex-1 min-w-0">{head}</div>
-        </div>
-        {expanded && rest}
-      </>
-    );
-  }
-
-  // Default Card now reads as a band — marker bar on the left, no enclosure.
-  // Variants:
-  //   glow    → highlighted band with primary marker
-  //   subtle  → quiet section with just internal padding & breathable spacing
-  //   frame   → opt-in enclosed frame for cases that genuinely need walls
-  if (variant === "frame") {
-    return <div className="ui-frame p-4 space-y-2.5">{body}</div>;
-  }
-  if (variant === "subtle") {
-    // Used heavily by plugins for choice grids — needs comfortable padding
-    // and inter-child spacing or the content reads as a wall of text.
-    return (
-      <div className="px-3 py-3 space-y-2 border-l-2 border-[var(--rule-color)] hover:border-[var(--accent-primary)] transition-colors">
-        {body}
-      </div>
-    );
-  }
-  return (
-    <div
-      className="ui-band space-y-2"
-      data-tone={variant === "glow" ? undefined : "muted"}
-    >
-      {body}
-    </div>
-  );
-};
-
-export const CardList: ComponentRenderer = ({ children }) => {
-  return <div className="space-y-2">{children}</div>;
-};
-
-export const EntryCard: ComponentRenderer = ({ element }) => {
-  const { t } = useTranslation();
-  const resolve = useI18nResolver();
-  const title = resolve(element.props?.title);
-  const category = (element.props?.category as string) ?? "";
-  const content = resolve(element.props?.content);
-  const tags = element.props?.tags as string[] | undefined;
-  const rarity = (element.props?.rarity as string) ?? "common";
-  // Optional plugin-supplied per-category icon + color (e.g. from a
-  // plugin's categoryMeta payload). When provided they override the
-  // built-in fallback map below; when missing the card still renders
-  // sensibly via the fallback so pre-enrichment entries keep working.
-  const externalIcon = element.props?.icon as string | undefined;
-  const externalColor = element.props?.color as string | undefined;
-  // Generic feature flags — any plugin can set these via itemLiteralProps
-  // or itemPropMap. collapsible toggles a chevron that hides body/tags;
-  // isNew renders a purple "NEW" sparkle next to the title.
-  const collapsible = (element.props?.collapsible as boolean) ?? false;
-  const defaultExpanded = (element.props?.defaultExpanded as boolean) ?? true;
-  const isNew = (element.props?.isNew as boolean) ?? false;
-
-  const [expanded, setExpanded] = useState(defaultExpanded);
-
-  const categoryIcons: Record<string, string> = {
-    monster: "skull",
-    item: "gem",
-    location: "map-pin",
-    lore: "scroll-text",
-    character: "users",
-    skill: "sparkles",
-  };
-  const rarityTone: Record<string, string> = {
-    legendary: "warning",
-    rare: "info",
-    uncommon: "info",
-    common: "muted",
-  };
-  const rarityMarkerColor: Record<string, string> = {
-    legendary: "var(--accent-warning)",
-    rare: "var(--accent-secondary)",
-    uncommon: "var(--accent-secondary)",
-    common: "var(--color-border)",
-  };
-  const rarityBadgeColors: Record<string, string> = {
-    legendary: "bg-amber-500/10 text-amber-600 border-amber-500/30",
-    rare: "bg-purple-500/10 text-purple-600 border-purple-500/30",
-    uncommon: "bg-blue-500/10 text-blue-600 border-blue-500/30",
-    common: "bg-muted text-muted-foreground border-border",
-  };
-  const categoryIconColors: Record<string, string> = {
-    red: "text-red-500 dark:text-red-400",
-    amber: "text-amber-500 dark:text-amber-400",
-    blue: "text-blue-500 dark:text-blue-400",
-    green: "text-green-500 dark:text-green-400",
-    purple: "text-purple-500 dark:text-purple-400",
-    cyan: "text-cyan-500 dark:text-cyan-400",
-  };
-
-  const CategoryIcon = resolveIcon(
-    externalIcon ?? categoryIcons[category] ?? "book-open",
-  );
-  const iconColorClass = externalColor
-    ? `${categoryIconColors[externalColor] ?? "text-primary"}`
-    : "text-primary";
-  const Chevron = expanded ? Icons.ChevronDown : Icons.ChevronRight;
-  const SparkleIcon = Icons.Sparkles;
-
-  const showBody = !collapsible || expanded;
-  const titleRowClass = clsx(
-    "flex items-center gap-2",
-    collapsible && "cursor-pointer select-none",
-  );
-
-  return (
-    <div
-      className="ui-band space-y-2"
-      data-tone={rarityTone[rarity] ?? "muted"}
-      style={{ ["--tw-band-marker" as string]: rarityMarkerColor[rarity] }}
-    >
-      <div
-        className={titleRowClass}
-        onClick={collapsible ? () => setExpanded((v) => !v) : undefined}
-        role={collapsible ? "button" : undefined}
-        tabIndex={collapsible ? 0 : undefined}
-        aria-expanded={collapsible ? expanded : undefined}
-        onKeyDown={
-          collapsible
-            ? (e) => {
-                if (e.key === "Enter" || e.key === " ") {
-                  e.preventDefault();
-                  setExpanded((v) => !v);
-                }
-              }
-            : undefined
-        }
-      >
-        {collapsible && (
-          <Chevron
-            className="w-3 h-3 shrink-0 text-muted-foreground"
-            aria-hidden="true"
-          />
-        )}
-        {CategoryIcon && (
-          <CategoryIcon
-            className={clsx("w-3.5 h-3.5 shrink-0", iconColorClass)}
-          />
-        )}
-        <span className="ui-entry-title text-[13px] font-medium flex-1 truncate text-foreground">
-          {title}
-        </span>
-        {isNew && (
-          <span
-            className="ui-chip inline-flex items-center gap-0.5 px-1.5 py-0.5 text-[9px] font-semibold tracking-wider uppercase text-purple-600 dark:text-purple-300 bg-purple-500/10 border border-purple-500/30"
-            aria-label={t("common.new", "new")}
-          >
-            <SparkleIcon className="w-2.5 h-2.5" />
-            {t("common.newUpper", "NEW")}
-          </span>
-        )}
-        <span
-          className={clsx(
-            "ui-chip inline-flex items-center px-1.5 py-0.5 text-[10px] border",
-            rarityBadgeColors[rarity],
-          )}
-        >
-          {category}
-        </span>
-      </div>
-      {showBody && content && (
-        <p className="text-[12.5px] text-muted-foreground leading-[1.6]">
-          {content}
-        </p>
-      )}
-      {showBody && tags && tags.length > 0 && (
-        <div className="flex flex-wrap gap-1">
-          {tags.map((tag) => (
-            <span
-              key={tag}
-              className="ui-chip text-[9px] px-1.5 py-0.5 bg-muted text-muted-foreground"
-            >
-              {tag}
-            </span>
-          ))}
-        </div>
-      )}
-    </div>
-  );
-};
 
 export const StatBar: ComponentRenderer = ({ element }) => {
   const resolve = useI18nResolver();
@@ -561,3 +343,6 @@ export const JsonView: ComponentRenderer = ({ element }) => {
   const value = element.props?.value;
   return <div className="text-[11px]">{renderJsonValue(value, 0)}</div>;
 };
+
+// ── Re-exports from sub-modules ────────────────────────────────────
+export { Card, CardList, EntryCard } from "./card-renderers.js";
